+++ src/main/java/com/vsu/amm/medframe/controller/ConfigurationController.java
@@ -2,8 +2,9 @@ package com.vsu.amm.medframe.controller;
 
 import com.vsu.amm.medframe.dto.DeviceDto;
 import com.vsu.amm.medframe.enums.Frequency;
-import com.vsu.amm.medframe.service.DevicePointService;
 import com.vsu.amm.medframe.service.DeviceService;
+import com.vsu.amm.medframe.service.impl.DevicePointServiceImpl;
+import com.vsu.amm.medframe.service.impl.DeviceServiceImpl;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -25,7 +26,7 @@ public class ConfigurationController {
     private DeviceService deviceService;
 
     @Autowired
-    private DevicePointService devicePointService;
+    private DevicePointServiceImpl devicePointService;
 
     @RequestMapping(value = "/devices/new", method = GET)
     public String getAddDeviceForm(ModelMap modelMap) {
