+++ src/test/java/com/vsu/medframe/test/service/DevicePointServiceTest.java
@@ -1,0 +1,4 @@
+package com.vsu.medframe.test.service;
+
+public class DevicePointServiceTest {
+}
