+++ src/test/java/com/vsu/medframe/test/service/DeviceServiceTest.java
@@ -1,0 +1,16 @@
+package com.vsu.medframe.test.service;
+
+import com.vsu.medframe.test.AbstractTest;
+import org.junit.Test;
+
+import static junit.framework.Assert.*;
+
+public class DeviceServiceTest extends AbstractTest {
+
+    @Test
+    public void testtest() {
+        int n = 4;
+        assertEquals(4, n);
+    }
+
+}
